requirements.txt
Django==1.6.4
South==0.8.4
django-shortuuidfield
django-tastypie
